+++ .eslintrc.js
@@ -14,7 +14,12 @@ const eslintConfig =
 		ecmaFeatures :
 		{
 			impliedStrict : true
-		}
+		},
+		lib : [ 'es2018' ]
+	},
+	globals :
+	{
+		NodeJS : 'readonly'
 	},
 	rules :
 	{
@@ -140,7 +145,7 @@ const eslintConfig =
 		'no-unmodified-loop-condition'  : 2,
 		'no-unreachable'                : 2,
 		'no-unused-vars'                : [ 1, { vars: 'all', args: 'after-used' } ],
-		'no-use-before-define'          : [ 2, { functions: false } ],
+		'no-use-before-define'          : 0,
 		'no-useless-call'               : 2,
 		'no-useless-computed-key'       : 2,
 		'no-useless-concat'             : 2,
@@ -211,7 +216,7 @@ switch (process.env.MEDIASOUP_NODE_LANGUAGE)
 					ignoreRestSiblings : false
 				}
 			],
-			'@typescript-eslint/no-use-before-define'  : 0,
+			'@typescript-eslint/no-use-before-define'  : [ 2, { functions: false } ],
 			'@typescript-eslint/no-empty-function'     : 0,
 			'@typescript-eslint/no-non-null-assertion' : 0
 		};