+++ src/Consumer.ts
@@ -391,6 +391,7 @@ export class Consumer extends EnhancedEventEmitter
 
 		// Remove notification subscriptions.
 		this._channel.removeAllListeners(this._internal.consumerId);
+		this._payloadChannel.removeAllListeners(this._internal.consumerId);
 
 		this._channel.request('consumer.close', this._internal)
 			.catch(() => {});