+++ src/PipeTransport.ts
@@ -20,6 +20,13 @@ export type PipeTransportOptions =
 	 */
 	listenIp: TransportListenIp | string;
 
+
+	/**
+	 * Disable IP:Port check for RTP and RTCP packets.
+	 * Default false.
+	 */
+	disableOriginCheck?: boolean;
+	
 	/**
 	 * Create a SCTP association. Default false.
 	 */
@@ -81,14 +88,13 @@ export type PipeTransportStat =
 	rtxRecvBitrate: number;
 	rtxBytesSent: number;
 	rtxSendBitrate: number;
-	probationBytesReceived: number;
-	probationRecvBitrate: number;
 	probationBytesSent: number;
 	probationSendBitrate: number;
 	availableOutgoingBitrate?: number;
 	availableIncomingBitrate?: number;
 	maxIncomingBitrate?: number;
 	// PipeTransport specific.
+	disableOriginCheck: boolean;
 	tuple: TransportTuple;
 }
 
@@ -99,6 +105,7 @@ export class PipeTransport extends Transport
 	// PipeTransport data.
 	protected readonly _data:
 	{
+		disableOriginCheck?: boolean;
 		tuple: TransportTuple;
 		sctpParameters?: SctpParameters;
 		sctpState?: SctpState;
@@ -121,11 +128,12 @@ export class PipeTransport extends Transport
 
 		this._data =
 		{
-			tuple          : data.tuple,
-			sctpParameters : data.sctpParameters,
-			sctpState      : data.sctpState,
-			rtx            : data.rtx,
-			srtpParameters : data.srtpParameters
+			disableOriginCheck : data.disableOriginCheck,
+			tuple              : data.tuple,
+			sctpParameters     : data.sctpParameters,
+			sctpState          : data.sctpState,
+			rtx                : data.rtx,
+			srtpParameters     : data.srtpParameters
 		};
 
 		this._handleWorkerNotifications();