+++ worker/src/RTC/PipeTransport.cpp
@@ -60,6 +60,16 @@ namespace RTC
 		if (jsonEnableRtxIt != data.end() && jsonEnableRtxIt->is_boolean())
 			this->rtx = jsonEnableRtxIt->get<bool>();
 
+		auto jsonDisableOriginCheckIt = data.find("disableOriginCheck");
+
+		if (jsonDisableOriginCheckIt != data.end())
+		{
+			if (!jsonDisableOriginCheckIt->is_boolean())
+				MS_THROW_TYPE_ERROR("wrong disableOriginCheck (not a boolean)");
+
+			this->disableOriginCheck = jsonDisableOriginCheckIt->get<bool>();
+		}
+
 		auto jsonEnableSrtpIt = data.find("enableSrtp");
 
 		// clang-format off
@@ -604,7 +614,7 @@ namespace RTC
 		}
 
 		// Verify that the packet's tuple matches our tuple.
-		if (!this->tuple->Compare(tuple))
+		if (!this->disableOriginCheck && !this->tuple->Compare(tuple))
 		{
 			MS_DEBUG_TAG(rtp, "ignoring RTP packet from unknown IP:port");
 
@@ -664,7 +674,7 @@ namespace RTC
 			return;
 
 		// Verify that the packet's tuple matches our tuple.
-		if (!this->tuple->Compare(tuple))
+		if (!this->disableOriginCheck && !this->tuple->Compare(tuple))
 		{
 			MS_DEBUG_TAG(sctp, "ignoring SCTP packet from unknown IP:port");